新增权限扫描

+++ clever-system/src/main/java/com/clever/service/impl/PermissionGroupServiceImpl.java
@@ -3,16 +3,23 @@ package com.clever.service.impl;
 import com.baomidou.mybatisplus.core.conditions.query.QueryWrapper;
 import com.baomidou.mybatisplus.extension.plugins.pagination.Page;
 import com.clever.bean.model.OnlineUser;
+import com.clever.bean.system.Permission;
+import com.clever.mapper.PermissionMapper;
+import com.clever.mapper.RolePermissionRelMapper;
+import com.clever.service.PermissionService;
+import com.clever.util.CollectionUtil;
 import org.apache.commons.lang3.StringUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.stream.Collectors;
 
 import com.clever.mapper.PermissionGroupMapper;
 import com.clever.bean.system.PermissionGroup;
 import com.clever.service.PermissionGroupService;
+import org.springframework.util.CollectionUtils;
 
 import javax.annotation.Resource;
 
@@ -30,6 +37,14 @@ public class PermissionGroupServiceImpl implements PermissionGroupService {
     @Resource
     private PermissionGroupMapper permissionGroupMapper;
 
+    @Resource
+    private RolePermissionRelMapper rolePermissionRelMapper;
+
+    @Resource
+    private PermissionMapper permissionMapper;
+    @Resource
+    private PermissionService permissionService;
+
     /**
      * 分页查询系统权限组列表
      *
@@ -42,9 +57,9 @@ public class PermissionGroupServiceImpl implements PermissionGroupService {
      * @return Page<PermissionGroup>
      */
     @Override
-    public Page<PermissionGroup> selectPage(Integer pageNumber, Integer pageSize, String platformId, String parentId, String name, Integer sortCode) {
+    public Page<PermissionGroup> selectPage(Integer pageNumber, Integer pageSize, Integer platformId, String parentId, String name, Integer sortCode) {
         QueryWrapper<PermissionGroup> queryWrapper = new QueryWrapper<>();
-        if (StringUtils.isNotBlank(platformId)) {
+        if (platformId != null) {
             queryWrapper.eq("platform_id", platformId);
         }
         if (StringUtils.isNotBlank(parentId)) {
@@ -77,7 +92,7 @@ public class PermissionGroupServiceImpl implements PermissionGroupService {
      * @return List<PermissionGroup> 系统权限组列表
      */
     @Override
-    public List<PermissionGroup> selectListByPlatformId(String platformId) {
+    public List<PermissionGroup> selectListByPlatformId(Integer platformId) {
         return permissionGroupMapper.selectList(new QueryWrapper<PermissionGroup>().eq("platform_id", platformId).orderByAsc("id"));
     }
 
@@ -91,6 +106,7 @@ public class PermissionGroupServiceImpl implements PermissionGroupService {
     public List<PermissionGroup> selectListByParentId(String parentId) {
         return permissionGroupMapper.selectList(new QueryWrapper<PermissionGroup>().eq("parent_id", parentId).orderByAsc("id"));
     }
+
 
     /**
      * 根据创建者id获取列表
@@ -166,6 +182,11 @@ public class PermissionGroupServiceImpl implements PermissionGroupService {
      */
     @Override
     public void deleteBatchIds(List<String> ids, OnlineUser onlineUser) {
+        if (ids == null || ids.isEmpty()) {
+            return;
+        }
+        rolePermissionRelMapper.deleteByPermissionGroupIds(ids);
+        permissionMapper.delete(new QueryWrapper<com.clever.bean.system.Permission>().in("group_id", ids));
         permissionGroupMapper.deleteBatchIds(ids);
         log.info("系统权限组, 系统权限组信息批量删除成功: userId={}, count={}, permissionGroupIds={}", onlineUser.getId(), ids.size(), ids.toString());
     }
@@ -177,7 +198,7 @@ public class PermissionGroupServiceImpl implements PermissionGroupService {
      * @param onlineUser 当前登录用户
      */
     @Override
-    public void deleteByPlatformId(String platformId, OnlineUser onlineUser) {
+    public void deleteByPlatformId(Integer platformId, OnlineUser onlineUser) {
         permissionGroupMapper.delete(new QueryWrapper<PermissionGroup>().eq("platform_id", platformId));
         log.info("系统权限组, 系统权限组信息根据platformId删除成功: userId={}, platformId={}", onlineUser.getId(), platformId);
     }
@@ -204,5 +225,70 @@ public class PermissionGroupServiceImpl implements PermissionGroupService {
     public void deleteByCreator(String creator, OnlineUser onlineUser) {
         permissionGroupMapper.delete(new QueryWrapper<PermissionGroup>().eq("creator", creator));
         log.info("系统权限组, 系统权限组信息根据creator删除成功: userId={}, creator={}", onlineUser.getId(), creator);
+    }
+
+    /**
+     * 解析需要删除的权限组
+     *
+     * @param platformId           平台id
+     * @param permissionGroupCodes 权限组id列表
+     */
+    public void resolvePermissionGroup(Integer platformId, List<String> permissionGroupCodes) {
+        List<String> oldPermissionGroupList = selectListByPlatformId(platformId).stream().map(PermissionGroup::getCode).collect(Collectors.toList());
+        CollectionUtil.compareAndRemove(permissionGroupCodes, oldPermissionGroupList);
+        if (!CollectionUtils.isEmpty(oldPermissionGroupList)) {
+            // 需要删除的权限组
+            OnlineUser onlineUser = new OnlineUser();
+            onlineUser.setId("system");
+            deleteBatchIds(oldPermissionGroupList, onlineUser);
+        }
+    }
+
+    /**
+     * 解析规则组添加
+     *
+     * @param platformId      平台id
+     * @param permissionGroup 规则组
+     * @param permissionList  规则列表
+     */
+    public void resolvePermissionGroup(Integer platformId, PermissionGroup permissionGroup, List<Permission> permissionList) {
+        PermissionGroup oldPermissionGroup = permissionGroupMapper.selectOne(new QueryWrapper<PermissionGroup>().eq("platform_id", platformId).eq("code", permissionGroup.getCode()));
+        if (oldPermissionGroup == null) {
+            // 新增
+            permissionGroup.setPlatformId(platformId);
+            create(permissionGroup, new OnlineUser());
+            permissionList.forEach(permission -> {
+                permission.setPlatformId(platformId);
+                permission.setGroupId(permissionGroup.getId());
+                permissionService.create(permission, new OnlineUser());
+            });
+        } else {
+            // 修改
+            permissionGroup.setId(oldPermissionGroup.getId());
+            update(permissionGroup, new OnlineUser());
+            List<String> oldPermissionList = permissionMapper.selectList(new QueryWrapper<Permission>().eq("platform_id", platformId).eq("group_id", oldPermissionGroup.getId())).stream().map(Permission::getCode).collect(Collectors.toList());
+            CollectionUtil.compareAndRemove(permissionList.stream().map(Permission::getCode).collect(Collectors.toList()), oldPermissionList);
+            if (!CollectionUtils.isEmpty(oldPermissionList)) {
+                // 需要删除的权限
+                OnlineUser onlineUser = new OnlineUser();
+                onlineUser.setId("system");
+                permissionService.deleteBatchIds(oldPermissionList, onlineUser);
+            }
+            permissionList.forEach(permission -> {
+                permission.setPlatformId(platformId);
+                permission.setGroupId(permissionGroup.getId());
+                permissionService.save(permission, new OnlineUser());
+            });
+        }
+    }
+
+    @Override
+    public PermissionGroup initPermissionGroup(PermissionGroup permissionGroup) {
+        PermissionGroup oldPermissionGroup = permissionGroupMapper.selectOne(new QueryWrapper<PermissionGroup>().eq("parent_id", permissionGroup.getParentId()).eq("platform_id", permissionGroup.getPlatformId()).eq("code", permissionGroup.getCode()));
+        if (oldPermissionGroup == null) {
+            permissionGroupMapper.insert(permissionGroup);
+            return permissionGroup;
+        }
+        return oldPermissionGroup;
     }
 }
